+++ crates/iota-types/Cargo.toml
@@ -60,7 +60,6 @@ tracing.workspace = true
 
 # internal dependencies
 consensus-config.workspace = true
-move-compiler.workspace = true
 iota-enum-compat-util.workspace = true
 iota-macros.workspace = true
 iota-network-stack.workspace = true
